META-253: Implement QnMetaDataV1::addMotion for QRectF (add 64 bit optimization)
META-253: Implement QnMetaDataV1::addMotion for QRectF (fast implementation)
META-253: Implement QnMetaDataV1::addMotion for QRectF (slow implementation)
META-253: Implement binary database for analytics for fast time periods earch
META-253: UT in progress
META-253: add 16 byte alignment to the analytics data
META-253: add cleanup unused data code
META-253: implement lookup timePeriods helper in the same way as motion
META-254: Analytics DB passes already translated region to the time period archive class
META-254: Analytics DB uses motion-like file for time periods
META-254: Analytics DB: some fixes in time periods tests
META-254: Analytics DB: switched to another QRectF -> QRect translation
META-254: nx::analytics::db::Filter::boundingBox is now optional
VMS-14108: add more log messages

<==== Mercurial data ====>
Branch: vms_4.0

+++ vms/server/nx_vms_server/src/nx/vms/server/analytics/db/attributes_dao.cpp
@@ -13,7 +13,7 @@ AttributesDao::AttributesDao()
     m_attributesCache.setMaxCost(kCacheSize);
 }
 
-long long AttributesDao::insertOrFetchAttributes(
+int64_t AttributesDao::insertOrFetchAttributes(
     sql::QueryContext* queryContext,
     const std::vector<common::metadata::Attribute>& eventAttributes)
 {
@@ -76,10 +76,10 @@ std::vector<common::metadata::Attribute> AttributesDao::deserialize(
 }
 
 void AttributesDao::addToAttributesCache(
-    long long id,
+    int64_t id,
     const QByteArray& content)
 {
-    auto cachedValue = std::make_unique<long long>(id);
+    auto cachedValue = std::make_unique<int64_t>(id);
     if (m_attributesCache.insert(
             QCryptographicHash::hash(content, QCryptographicHash::Md5),
             cachedValue.get()))
@@ -88,12 +88,12 @@ void AttributesDao::addToAttributesCache(
     }
 }
 
-long long AttributesDao::findAttributesIdInCache(
+int64_t AttributesDao::findAttributesIdInCache(
     const QByteArray& content)
 {
     const auto md5 = QCryptographicHash::hash(content, QCryptographicHash::Md5);
 
-    long long* id = m_attributesCache.object(md5);
+    int64_t* id = m_attributesCache.object(md5);
     if (id == nullptr)
         return -1;
     return *id;
